+++ parser_library/src/config/proc_grps.h
@@ -42,6 +42,8 @@ struct db2_preprocessor
     bool valid() const noexcept { return version.size() <= 64; }
 
     friend bool operator==(const db2_preprocessor&, const db2_preprocessor&) = default;
+
+    constexpr static std::string_view name = "DB2";
 };
 
 void to_json(nlohmann::json& j, const db2_preprocessor& p);
@@ -56,14 +58,28 @@ struct cics_preprocessor
     bool valid() const noexcept { return true; }
 
     friend bool operator==(const cics_preprocessor&, const cics_preprocessor&) = default;
+
+    constexpr static std::string_view name = "CICS";
 };
 
 void to_json(nlohmann::json& j, const cics_preprocessor& p);
 void from_json(const nlohmann::json& j, cics_preprocessor& p);
 
+struct endevor_preprocessor
+{
+    bool valid() const noexcept { return true; }
+
+    friend bool operator==(const endevor_preprocessor&, const endevor_preprocessor&) = default;
+
+    constexpr static std::string_view name = "ENDEVOR";
+};
+
+void to_json(nlohmann::json& j, const endevor_preprocessor& p);
+void from_json(const nlohmann::json& j, endevor_preprocessor& p);
+
 struct preprocessor_options
 {
-    std::variant<db2_preprocessor, cics_preprocessor> options;
+    std::variant<db2_preprocessor, cics_preprocessor, endevor_preprocessor> options;
 
     bool valid() const noexcept;