+++ parser_library/src/diagnostic.h
@@ -709,6 +709,10 @@ struct diagnostic_op
 
     static diagnostic_op warn_CIC002(const range& range, std::string_view variable_name);
 
+    static diagnostic_op error_END001(const range& range, std::string_view lib);
+
+    static diagnostic_op error_END002(const range& range, std::string_view lib);
+
     static diagnostic_op warn_U001_drop_had_no_effect(const range& range, std::string_view arg);
     static diagnostic_op warn_U001_drop_had_no_effect(const range& range, int);