+++ parser_library/src/preprocessor_options.h
@@ -49,7 +49,13 @@ struct cics_preprocessor_options
     bool operator==(const cics_preprocessor_options&) const = default;
 };
 
-using preprocessor_options = std::variant<db2_preprocessor_options, cics_preprocessor_options>;
+struct endevor_preprocessor_options
+{
+    bool operator==(const endevor_preprocessor_options&) const = default;
+};
+
+using preprocessor_options =
+    std::variant<db2_preprocessor_options, cics_preprocessor_options, endevor_preprocessor_options>;
 
 } // namespace hlasm_plugin::parser_library
 #endif