+++ parser_library/test/processing/multiple_preprocessors_test.cpp
@@ -14,6 +14,7 @@
 
 
 #include "../common_testing.h"
+#include "../mock_parse_lib_provider.h"
 
 TEST(multiple_preprocessors, cics_and_db2)
 {
@@ -33,4 +34,33 @@ RO       SQL  TYPE IS ROWID
     a.collect_diags();
 
     EXPECT_TRUE(a.diags().empty());
+}
+
+TEST(multiple_preprocessors, endevor_and_cics)
+{
+    mock_parse_lib_provider libs({
+        { "MEMBER", R"(
+         LARL 0,DFHVALUE(FIRSTQUIESCE)
+)" },
+    });
+    std::string input = R"(
+TEST    DS     0C
+-INC MEMBER
+        LTORG
+TESTLEN EQU    *-TEST
+        END
+)";
+
+    analyzer a(input,
+        analyzer_options { &libs,
+            std::vector<preprocessor_options> {
+                preprocessor_options(endevor_preprocessor_options()),
+                preprocessor_options(cics_preprocessor_options(false, false, false)),
+            } });
+    a.analyze();
+    a.collect_diags();
+
+    EXPECT_TRUE(a.diags().empty());
+
+    EXPECT_EQ(get_symbol_abs(a.hlasm_ctx(), "TESTLEN"), 12);
 }